requirements.txt
aioredis==2.0.1
aiosmtplib==1.1.6
alembic==1.7.7
anyio==3.5.0
asgiref==3.5.0
async-timeout==4.0.2
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Deprecated==1.2.13
distlib==0.3.4
dnspython==2.2.1
email-validator==1.1.3
fakeredis==1.7.1
fastapi==0.75.1
fastapi-mail==1.0.8
FastAPI-SQLAlchemy==0.2.1
filelock==3.6.0
greenlet==1.1.2
h11==0.12.0
httpcore==0.14.7
httptools==0.4.0
httpx==0.22.0
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.6.0
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
packaging==21.3
platformdirs==2.5.1
psycopg2==2.9.3
pydantic==1.9.0
PyJWT==2.3.0
pyparsing==3.0.8
python-dotenv==0.20.0
python-multipart==0.0.5
PyYAML==6.0
redis==4.1.4
requests==2.27.1
rfc3986==1.5.0
schedule==1.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.35
starlette==0.17.1
typing-extensions==4.1.1
unicorn==1.0.3
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
virtualenv==20.14.0
watchgod==0.8.2
websockets==10.2
wrapt==1.14.0
zipp==3.8.0
